requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy<2
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
nltk>=3.8.0
scikit-image>=0.20.0
pycocoevalcap==1.2
